added user authorization

+++ src/app/controllers/dm_controllers/delete_dm_group_controller.py
@@ -7,14 +7,17 @@ class DeleteDmGroupController(Controller):
         return "delete-dm-group"
 
     def get_path(self):
-        return "/dms/delete/"
+        return "/dms/<int:dm_group_id>/"
     
     def get_methods(self):
         return ["DELETE"]
     
-    def content(self):
-        data = request.get_json()
-        dm_id = data.get("dm_group_id")
+    @authorize_user
+    def content(self, **kwargs):
+        user = kwargs.get("user")
+        dm_group = dms_dao.get_dm_group_by_id(dm_group_id)
+        if user not in dm_group.users:
+            raise Exception("User not in DM Group")
         dm_group = dms_dao.delete_dm_group_by_id(dm_id)
         return dm_group.serialize()
