+++ src/app/controllers/workspace_controllers/delete_workspace_controller.py
@@ -7,14 +7,18 @@ class DeleteWorkspaceController(Controller):
         return "delete-workspace"
 
     def get_path(self):
-        return "/workspaces/delete/"
+        return "/workspaces/<int:workspace_id>/"
     
     def get_methods(self):
         return ["DELETE"]
     
-    def content(self):
-        data = request.get_json()
-        workspace_id = data.get("workspace_id")
+    @authorize_user
+    def content(self, **kwargs):
+        user = kwargs.get("user")
+        in_workspace, _ = workspaces_dao.is_user_in_workspace(user.id, workspace_id)
+        if not in_workspace:
+            raise Exception("User not in workspace")
+
         workspace = workspaces_dao.delete_workspace_by_id(workspace_id)
         return workspace.serialize()
